export helper functions from util file
